# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "description": "새로운 조직을 생성합니다.",
        "operationId": "OrganizationController_createOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "조직 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "테스트 학교"
                    },
                    "type": {
                      "type": "string",
                      "example": "school"
                    },
                    "mainAdminRole": {
                      "type": "string",
                      "example": "MAIN_ADMIN"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 생성",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}": {
      "get": {
        "description": "특정 조직의 상세 정보를 조회합니다.",
        "operationId": "OrganizationController_getOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조직 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "테스트 학교"
                    },
                    "type": {
                      "type": "string",
                      "example": "school"
                    },
                    "mainAdminRole": {
                      "type": "string",
                      "example": "MAIN_ADMIN"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "조직을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 상세 조회",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/members": {
      "get": {
        "description": "조직의 모든 멤버를 조회합니다.",
        "operationId": "OrganizationController_getOrganizationMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멤버 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "profileId": {
                        "type": "number",
                        "example": 1
                      },
                      "roleInOrg": {
                        "type": "string",
                        "example": "STUDENT"
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "홍길동"
                          },
                          "role": {
                            "type": "string",
                            "example": "STUDENT"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 멤버 목록 조회",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "description": "조직에 새로운 멤버를 추가합니다.",
        "operationId": "OrganizationController_addMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "멤버 추가 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "멤버가 성공적으로 추가되었습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "프로필을 찾을 수 없음"
          },
          "409": {
            "description": "이미 조직에 존재하는 멤버"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 멤버 추가",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/members/{role}": {
      "get": {
        "description": "특정 역할을 가진 조직 멤버들을 조회합니다.",
        "operationId": "OrganizationController_getOrganizationMembersByRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "path",
            "description": "조회할 역할",
            "schema": {
              "enum": [
                "MAIN_ADMIN",
                "SUB_ADMIN",
                "STUDENT",
                "PARENT"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "역할별 멤버 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "profileId": {
                        "type": "number",
                        "example": 1
                      },
                      "roleInOrg": {
                        "type": "string",
                        "example": "STUDENT"
                      },
                      "profile": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "홍길동"
                          },
                          "role": {
                            "type": "string",
                            "example": "STUDENT"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "역할별 조직 멤버 조회",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/members/{profileId}/role": {
      "put": {
        "description": "조직 멤버의 역할을 변경합니다.",
        "operationId": "OrganizationController_updateMemberRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "역할 변경 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "역할이 성공적으로 변경되었습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "멤버를 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "멤버 역할 변경",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/members/{profileId}": {
      "delete": {
        "description": "조직에서 멤버를 제거합니다.",
        "operationId": "OrganizationController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조직 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멤버 제거 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "멤버가 성공적으로 제거되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "멤버를 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "조직 멤버 제거",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/my-organizations": {
      "get": {
        "description": "현재 사용자가 속한 모든 조직을 조회합니다.",
        "operationId": "OrganizationController_getMyOrganizations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "조직 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "테스트 학교"
                      },
                      "type": {
                        "type": "string",
                        "example": "school"
                      },
                      "roleInOrg": {
                        "type": "string",
                        "example": "STUDENT"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "내 조직 목록 조회",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/my-admin-organizations": {
      "get": {
        "description": "현재 사용자가 주 관리자인 조직들을 조회합니다.",
        "operationId": "OrganizationController_getMyAdminOrganizations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "관리 조직 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "테스트 학교"
                      },
                      "type": {
                        "type": "string",
                        "example": "school"
                      },
                      "roleInOrg": {
                        "type": "string",
                        "example": "MAIN_ADMIN"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "내 관리 조직 목록 조회",
        "tags": [
          "organizations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Thinking API Server",
    "description": "수업 관리 시스템 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateOrganizationPayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "조직명",
            "example": "테스트 학교"
          },
          "type": {
            "type": "string",
            "description": "조직 유형",
            "example": "school"
          },
          "mainAdminRole": {
            "type": "string",
            "description": "주 관리자 역할",
            "enum": [
              "MAIN_ADMIN",
              "SUB_ADMIN",
              "STUDENT",
              "PARENT"
            ],
            "example": "MAIN_ADMIN"
          }
        },
        "required": [
          "name",
          "type",
          "mainAdminRole"
        ]
      },
      "AddMemberPayload": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "number",
            "description": "추가할 멤버의 프로필 ID",
            "example": 1
          },
          "roleInOrg": {
            "type": "string",
            "description": "조직 내 역할",
            "enum": [
              "MAIN_ADMIN",
              "SUB_ADMIN",
              "STUDENT",
              "PARENT"
            ],
            "example": "STUDENT"
          }
        },
        "required": [
          "profileId",
          "roleInOrg"
        ]
      },
...
    }
  }
}